Move lowerDash into utilities group
